refactored template services and routes
